+++ src/PlaywrightSharp/Transport/Channels/PageChannel.cs
@@ -328,7 +328,7 @@ namespace PlaywrightSharp.Transport.Channels
                     ["clickCount"] = clickCount,
                 });
 
-        internal Task MouseMoveAsync(double x, double y, int? steps)
+        internal Task MouseMoveAsync(decimal x, decimal y, int? steps)
             => Scope.SendMessageToServer(
                 Guid,
                 "mouseMove",
@@ -349,7 +349,7 @@ namespace PlaywrightSharp.Transport.Channels
                     ["clickCount"] = clickCount,
                 });
 
-        internal Task MouseClickAsync(double x, double y, int delay, MouseButton button, int clickCount)
+        internal Task MouseClickAsync(decimal x, decimal y, int delay, MouseButton button, int clickCount)
             => Scope.SendMessageToServer(
                 Guid,
                 "mouseClick",
